ai_principal/cli/requirements.txt
tabulate>=0.9.0
rich>=13.0.0
pandas>=2.1.1
python-dotenv==1.0.0